BLOXORZ_GAME/requirements.txt
numpy==1.23.0
pyOpenGL==3.1.6
PyOpenGL_accelerate==3.1.6
pygame==2.1.2
opencv-python==4.6.0.66
